requirements.txt
aniso8601==6.0.0
appdirs==1.4.3
argon2-cffi==19.1.0
attrs==19.1.0
black==19.3b0
certifi==2019.3.9
cffi==1.12.3
Click==7.0
cloudinary==1.15.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Extended==3.18.1
flask-marshmallow==0.10.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==3.0.0rc5
marshmallow-sqlalchemy==0.16.1
mock==2.0.0
pbr==5.1.3
Pillow==6.0.0
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.10.1
python-http-client==3.1.0
pytz==2018.9
sendgrid==6.0.4
six==1.12.0
SQLAlchemy==1.3.2
toml==0.10.0
urllib3==1.24.2
Werkzeug==0.15.2
